requirements.txt
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
django-debug-toolbar==1.10.1
python-decouple==3.1
